requirements.txt
bcrypt==3.2.2
blinker==1.4
certifi==2022.6.15
cffi==1.15.0
click==8.0.4
colorama==0.4.4
dnspython==2.2.1
email-validator==1.2.1
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.1
Pillow==9.1.1
pip==22.1
psycopg2==2.9.3
pycparser==2.21
python-dotenv==0.20.0
setuptools==61.2.0
SQLAlchemy==1.4.39
Werkzeug==2.0.3
wheel==0.37.1
wincertstore==0.2
WTForms==3.0.1
